requirements.txt
altair==4.2.0
geopandas==0.11.1
numpy==1.23.3
pandas==1.4.4
protobuf
psycopg2-binary
pydeck==0.7.1
pygsheets==2.0.5
Shapely==1.8.4
scipy==1.9.1
